Refactor error crate
